texture_atlas made larger

+++ source/block/block_cast_block.c
@@ -35,19 +35,19 @@ getSideMask(struct block_info* this, enum side side, struct block_info* it) {
 	return face_occlusion_full();
 }
 
-static uint8_t getTextureIndex1(struct block_info* this, enum side side) {
+static uint16_t getTextureIndex1(struct block_info* this, enum side side) {
 	return tex_atlas_lookup(TEXAT_CAST_BLOCK_IRON);
 }
 
-static uint8_t getTextureIndex2(struct block_info* this, enum side side) {
+static uint16_t getTextureIndex2(struct block_info* this, enum side side) {
 	return tex_atlas_lookup(TEXAT_CAST_BLOCK_GOLD);
 }
 
-static uint8_t getTextureIndex3(struct block_info* this, enum side side) {
+static uint16_t getTextureIndex3(struct block_info* this, enum side side) {
 	return tex_atlas_lookup(TEXAT_CAST_BLOCK_DIAMOND);
 }
 
-static uint8_t getTextureIndex4(struct block_info* this, enum side side) {
+static uint16_t getTextureIndex4(struct block_info* this, enum side side) {
 	return tex_atlas_lookup(TEXAT_CAST_BLOCK_LAPIS);
 }
