+++ source/block/block_lava.c
@@ -48,7 +48,7 @@ getSideMask(struct block_info* this, enum side side, struct block_info* it) {
 	#endif
 }
 
-static uint8_t getTextureIndex(struct block_info* this, enum side side) {
+static uint16_t getTextureIndex(struct block_info* this, enum side side) {
 	#ifdef GFX_FANCY_LIQUIDS
 	return TEXTURE_INDEX(2, 0);
 	#else
