+++ source/block/block_leaves.c
@@ -53,7 +53,7 @@ getSideMask(struct block_info* this, enum side side, struct block_info* it) {
 //	return face_occlusion_full();
 }
 
-static uint8_t getTextureIndex(struct block_info* this, enum side side) {
+static uint16_t getTextureIndex(struct block_info* this, enum side side) {
 	switch(this->block->metadata & 0x3) {
 		default:
 		case 0: return tex_atlas_lookup(TEXAT_LEAVES_OAK);
