+++ source/block/block_sandstone_stairs.c
@@ -113,7 +113,7 @@ getSideMask(struct block_info* this, enum side side, struct block_info* it) {
 	}
 }
 
-static uint8_t getTextureIndex(struct block_info* this, enum side side) {
+static uint16_t getTextureIndex(struct block_info* this, enum side side) {
 	(void)this;
 	switch(side) {
 		case SIDE_TOP: return tex_atlas_lookup(TEXAT_SANDSTONE_TOP);
