+++ source/block/blocks.h
@@ -35,7 +35,7 @@
 struct block {
 	char name[32];
 	enum block_material (*getMaterial)(struct block_info*);
-	uint8_t (*getTextureIndex)(struct block_info*, enum side);
+	uint16_t (*getTextureIndex)(struct block_info*, enum side);
 	struct face_occlusion* (*getSideMask)(struct block_info*, enum side,
 										  struct block_info*);
 	size_t (*getBoundingBox)(struct block_info*, bool, struct AABB*);
